Refactor Profile page - connect vs connected states ; Add total Claimed summuary
