+++ hooks/useClaims.ts
@@ -7,13 +7,15 @@ import {
   getClaimWeeks,
   getUnclaimedWeeksForAddress,
   getUnclaimedWeeksValues,
-  getUnclaimedTotal,
+  getClaimedWeeksValues,
+  getWeekValuesTotal,
   getClaimsWeeksProofs,
 } from "@/utils/claims";
 import type { ClaimWeek } from "@/utils/claims";
 
 const useClaims = () => {
   const [unclaimedTotal, setUnclaimedTotal] = useState(0);
+  const [claimedTotal, setClaimedTotal] = useState(0);
   const { loading, account, provider } = useWeb3Context();
   useEffect(() => {
     const fetchClaims = async () => {
@@ -26,17 +28,18 @@ const useClaims = () => {
       const redeem = MerkleRedeem__factory.connect(MERKLE_REDEEM_CONTRACT, provider.getSigner());
       const unclaimedWeeks = await getUnclaimedWeeksForAddress(redeem, claimWeeks, address);
       const unclaimedWeeksValues = getUnclaimedWeeksValues(claimWeeks, unclaimedWeeks, address);
+      const claimedWeeksValues = getClaimedWeeksValues(claimWeeks, unclaimedWeeks, address);
       const claimWeeksProofs = getClaimsWeeksProofs(claimWeeks, unclaimedWeeksValues, address);
       const result = await redeem.verifyClaim(address, claimWeeksProofs[0][0], claimWeeksProofs[0][1], claimWeeksProofs[0][2]);
 
-      setUnclaimedTotal(getUnclaimedTotal(unclaimedWeeksValues));
-      console.log(result);
+      setUnclaimedTotal(getWeekValuesTotal(unclaimedWeeksValues));
+      setClaimedTotal(getWeekValuesTotal(claimedWeeksValues));
     };
 
     fetchClaims();
   }, [account, loading, provider]);
 
-  return { unclaimedTotal };
+  return { unclaimedTotal, claimedTotal };
 };
 
 export default useClaims;
