--- pages/connect.tsx
@@ -1,3 +1,0 @@
-import Connect from "@/components/connect";
-
-export default Connect;
