+++ pages/profile.tsx
@@ -1,0 +1,3 @@
+import Profile from "@/components/profile";
+
+export default Profile;
